requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
celery[redis]
langchain
langchain-groq
langgraph
llama-index
llama-index-vector-stores-chroma
chromadb
mlflow
httpx
python-multipart
pyyaml
